refactor: FSD 구조에 맞춰 투표 관련 컴포넌트 재구성
- features/voting: 투표 비즈니스 로직 컴포넌트 이동
- widgets/voting: 투표 모달 및 헤더 위젯 이동
- shared/resetButton: 범용 리셋 버튼 컴포넌트 이동
