+++ frontend/src/features/voting/components/voteCountIcons/VoteCountIcons.tsx
