+++ frontend/src/features/voting/components/voteCountIcons/voteCountIcons.styled.ts
