+++ frontend/src/features/voting/components/voteOption/VoteOption.tsx
