+++ frontend/src/features/voting/components/voteOption/voteOption.styled.ts
