+++ frontend/src/features/voting/components/voteOptions/voteOptions.styled.ts
