+++ frontend/src/shared/components/ResetButton/resetButton.styled.ts
