+++ frontend/src/widgets/voting/components/voteHeader/VoteHeader.tsx
