+++ frontend/src/widgets/voting/components/voteHeader/voteHeader.styled.ts
