+++ frontend/src/widgets/voting/components/voteModal/voteModal.styled.ts
