+++ frontend/src/widgets/voting/hooks/useVoteModal.tsx
@@ -1,6 +1,6 @@
 import { modal } from '@features/modal/utils/modal';
 
-import VoteModal from '../components/VoteModal';
+import VoteModal from '../components/voteModal/VoteModal';
 
 const useVoteModal = () => {
   const openVoteModal = () => {
@@ -24,3 +24,4 @@ const useVoteModal = () => {
 };
 
 export default useVoteModal;
+
